Geo_Tour-main/requirements.txt
openai>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
replicate>=0.25.0
pillow>=10.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
google-generativeai>=0.8.0
playwright>=1.40.0
